+++ src/api.js
@@ -18,39 +18,44 @@ class API {
     if (!this.isInTestMode) {
       // Check permissions
       this.checkPermissions()
+        .then(e => {
+          log.info(e)
+        })
+        .catch(e => {
+          log.warn(e)
+        })
 
       // Start the queue processing
       setInterval(() => {
         this.processQueue()
+          .catch(e => {
+            log.warn(e)
+          })
       }, 1000)
     }
   }
 
-  checkPermissions () {
+  async checkPermissions () {
     // Check if the token has the required permissions
-    this.scheduleCall('groups.getTokenPermissions', {}, json => {
-      if (!json.response) {
-        log.error(`While checking token permission "messages", an error occured: ${json.error}`)
-      }
+    var response = await this.scheduleCall('groups.getTokenPermissions', {})
 
-      var permissions = json.response.permissions
-      var ok = false
-      for (var permission of permissions) {
-        if (permission.name === 'messages') {
-          ok = true
-          break
-        }
+    var permissions = response.permissions
+    var ok = false
+    for (var permission of permissions) {
+      if (permission.name === 'messages') {
+        ok = true
+        break
       }
+    }
 
-      if (!ok) {
-        log.warn('Token permission "messages" is missing. Bot will be unable to send any messages.')
-      } else {
-        log.info('Token permission "messages" is present.')
-      }
-    })
+    if (!ok) {
+      return Promise.reject(new Error('Token permission "messages" is missing. Bot will be unable to send any messages.'))
+    } else {
+      return Promise.resolve('Token permission "messages" is present.')
+    }
   }
 
-  processQueue () {
+  async processQueue () {
     if (this.queue) {
       for (var i = 1; i <= this.API_QUOTA; i++) {
         if (this.queue.length === 0) {
@@ -59,25 +64,40 @@ class API {
 
         var e = this.queue.shift()
 
-        this.call(e.method, e.params)
-          .then((json) => {
-            if (e.callback) {
-              e.callback(json)
-            }
-          })
+        var json = await this.call(e.method, e.params)
+
+        if (json.response !== undefined && json.response !== null) {
+          e.resolve(json.response)
+        } else {
+          if (json.error) {
+            var errorCode = json.error.error_code
+            var errorMsg = json.error.error_msg
+
+            e.reject(`An API call to method '${e.method}' failed due to an API error #${errorCode}: ${errorMsg}`)
+          } else {
+            e.reject(`An API call to method '${e.method}' failed due to an unknown API error. The API responded with: ${JSON.stringify(json)}`)
+          }
+        }
       }
+
+      return Promise.resolve()
     }
+
+    return Promise.reject(new Error('No queue for API calls found'))
   }
 
-  scheduleCall (method, params, callback) {
-    this.queue.push({
-      method: method,
-      params: params,
-      callback: callback
+  async scheduleCall (method, params) {
+    return new Promise((resolve, reject) => {
+      this.queue.push({
+        method: method,
+        params: params,
+        resolve: resolve,
+        reject: reject
+      })
     })
   }
 
-  call (method, params) {
+  async call (method, params) {
     method = encodeURIComponent(method)
     var url = `https://api.vk.com/method/${method}`
 
@@ -101,26 +121,23 @@ class API {
     return promise
   }
 
-  send (pid, msg, attachment) {
+  async send (pid, msg, attachment) {
     var params = {
       peer_id: pid,
       message: msg,
       attachment: attachment
     }
 
-    this.scheduleCall('messages.send', params, (json) => {
-      if (json.response) {
-        this.stats.sent()
-      } else {
-        if (json.error) {
-          var errorCode = json.error.error_code
-          var errorMsg = json.error.error_msg
-
-          log.warn(`A message was not sent to peer ${pid} due to an API error #${errorCode}: ${errorMsg}`)
-        } else {
-          log.warn(`A message was not sent to peer ${pid} due to an unknown API error. The API responded with: ${json.toString()}`)
-        }
-      }
+    return new Promise((resolve, reject) => {
+      this.scheduleCall('messages.send', params)
+        .then(_ => {
+          this.stats.sent()
+          resolve()
+        })
+        .catch(e => {
+          log.warn(e)
+          resolve()
+        })
     })
   }
 }